fixed weird spacing in page content
