+++ src/app/(portal)/portal/directory/page.tsx
@@ -4,8 +4,10 @@ import ProjectsDirectory from "@/components/portal/directory/ProjectsDirectory";
 
 export default function PortalDirectory() {
     return (
-        <div className="p-8 text-black">
+        <div className="flex-1 text-black px-6 py-8 md:px-10 md:py-10">
+            <div className="mx-auto max-w-6xl">
                 <ProjectsDirectory />
+            </div>
         </div>
     );
 }
