+++ src/app/(portal)/portal/layout.tsx
@@ -5,9 +5,9 @@ import MembershipPortalNavbar from '@/components/MembershipPortalNavbar';
 
 const portalTabs = [
     { id: 'dashboard', label: 'Dashboard', href: '/portal' },
-    { id: 'project-directory', label: 'Project Directory', href: '/portal/projects' },
+    { id: 'project-directory', label: 'Project Directory', href: '/portal/directory' },
     { id: 'my-project', label: 'My Project', href: '/portal/projects/overview' },
-    { id: 'member-directory', label: 'Member Directory', href: '/portal/directory' },
+    { id: 'member-directory', label: 'Member Directory', href: '/portal/members-directory' },
     { id: 'profile', label: 'Profile', href: '/portal/profile' },
 ];
 
@@ -17,9 +17,9 @@ export default function PortalPageLayout({
     children: React.ReactNode;
 }) {
     return (
-        <div className="min-h-screen">
+        <div className="min-h-screen flex flex-col">
             <MembershipPortalNavbar tabs={portalTabs} />
-            <main>{children}</main>
+            <main className="flex flex-1 bg-[#F6F8FA]">{children}</main>
         </div>
     );
 }
