+++ src/app/(portal)/portal/profile/page.tsx
@@ -3,8 +3,10 @@
 
 export default function PortalProfile() {
     return (
-        <div className="p-8 text-black">
-            <h1 className="text-3xl font-bold mb-4">Profile</h1>
+        <div className="flex-1 text-black px-6 py-8 md:px-10 md:py-10">
+            <div className="mx-auto max-w-6xl">
+                <h1 className="text-3xl md:text-4xl font-semibold tracking-tight">Profile</h1>
+            </div>
         </div>
     );
 }
