--- src/app/(portal)/portal/projects/board/page.tsx
@@ -1,26 +1,0 @@
-"use client";
-
-import MembershipPortalSidebar from "@/components/MembershipPortalSidebar";
-import ProjectBoardContent from "@/components/ProjectBoardContent";
-
-export default function Page() {
-  return (
-    <div className="flex flex-1 bg-[#F6F8FA] text-black overflow-x-hidden">
-        {/* Left sidebar */}
-        <MembershipPortalSidebar className="flex-shrink-0 sticky top-24" />
-
-        {/* Right content area */}
-        <div className="flex-1 px-6 py-10 md:px-16 md:py-12">
-          <div className="mx-auto max-w-7xl space-y-6">
-            <section className="space-y-2">
-              <h1 className="text-3xl md:text-4xl font-semibold tracking-tight">
-                Board
-              </h1>
-            </section>
-
-            <ProjectBoardContent />
-          </div>
-        </div>
-    </div>
-  );
-}
