+++ src/components/portal/directory/ProjectsDirectory.css
@@ -1,15 +1,3 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f5f5f5;
-    padding: 20px;
-}
-
 .container {
     max-width: 1200px;
     margin: 0 auto;
